refactor: reorg the files
