+++ apps/extension/src/entrypoints/options/pages/translation/personalized-prompt/utils/prompt-file.ts
